+++ NEWS.md
@@ -1,6 +1,11 @@
 ## version 2.0.1
 ---
 
+### new features
+- `ref_level` can be specified in `get_mob_stats` which allows for the 
+reference (i.e., control) group to be specified which all others are compared to.
+This is similar to the implementation in the function `get_delta_stats`
+
 ### minor bug fixes
 - typo in test_rarefaction.R
 - typo in get_mob_stats documentation